requirements.txt
tensorflow
keras
matplotlib
numpy
pandas
mlflow
kaggle
pytest
pylint
opentelemetry-api 
opentelemetry-sdk
opentelemetry-instrumentation
opentelemetry-distro